Don't generate a function name in `magit-define-command'
Make `magit-define-command` define a function whose name literally
matches the name specified in the call, instead of prepending `magit-`
to it.  Also modify all the calls to `magit-define-command` to include
the `magit-` prefix themselves.

The advantage is that when calling describe-function on the generated
command, the *Help* buffer will now include a link to the line in
`magit.el` where the command is defined.  Previously it would be
unable to find the definition, because the definition did not include
the function's real name.